EIP1-5345 Refactor the logging to be in separate package for reuse.
