+++ src/test/kotlin/uk/gov/dluhc/logging/testsupport/assertj/assertions/ILoggingEventAssert.kt
@@ -1,9 +1,9 @@
-package uk.gov.dluhc.printapi.testsupport.assertj.assertions
+package uk.gov.dluhc.logging.testsupport.assertj.assertions
 
 import ch.qos.logback.classic.spi.ILoggingEvent
-import org.apache.commons.lang3.StringUtils.isBlank
+import org.apache.commons.lang3.StringUtils
 import org.assertj.core.api.AbstractAssert
-import uk.gov.dluhc.printapi.config.CORRELATION_ID
+import uk.gov.dluhc.logging.config.CORRELATION_ID
 
 class ILoggingEventAssert(actual: ILoggingEvent?) :
     AbstractAssert<ILoggingEventAssert, ILoggingEvent?>(actual, ILoggingEventAssert::class.java) {
@@ -25,7 +25,7 @@ class ILoggingEventAssert(actual: ILoggingEvent?) :
     fun hasAnyCorrelationId(): ILoggingEventAssert {
         isNotNull
         with(actual!!) {
-            if (isBlank(mdcPropertyMap[CORRELATION_ID])) {
+            if (StringUtils.isBlank(mdcPropertyMap[CORRELATION_ID])) {
                 failWithMessage("Expected log message to have a correlation ID, but it did not")
             }
         }
